@@ -10,8 +10,10 @@
 		"email": "sindresorhus@gmail.com",
 		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=10"
+		"node": "^12.20.0 || ^14.13.1 || >=16.0.0"
 	},
 	"scripts": {
 		"test": "xo && ava && tsd"
@@ -37,22 +39,22 @@
 		"idle"
 	],
 	"dependencies": {
-		"bl": "^4.1.0",
-		"chalk": "^4.1.0",
-		"cli-cursor": "^3.1.0",
-		"cli-spinners": "^2.5.0",
-		"is-interactive": "^1.0.0",
-		"is-unicode-supported": "^0.1.0",
-		"log-symbols": "^4.1.0",
-		"strip-ansi": "^6.0.0",
+		"bl": "^5.0.0",
+		"chalk": "^4.1.2",
+		"cli-cursor": "^4.0.0",
+		"cli-spinners": "^2.6.0",
+		"is-interactive": "^2.0.0",
+		"is-unicode-supported": "^1.1.0",
+		"log-symbols": "^5.0.0",
+		"strip-ansi": "^7.0.0",
 		"wcwidth": "^1.0.1"
 	},
 	"devDependencies": {
-		"@types/node": "^14.14.35",
-		"ava": "^2.4.0",
-		"get-stream": "^6.0.0",
+		"@types/node": "^16.7.1",
+		"ava": "^3.15.0",
+		"get-stream": "^6.0.1",
 		"transform-tty": "^1.0.11",
-		"tsd": "^0.14.0",
-		"xo": "^0.38.2"
+		"tsd": "^0.17.0",
+		"xo": "^0.44.0"
 	}
 }
