@@ -1,7 +1,6 @@
+import {PassThrough as PassThroughStream} from 'node:stream';
 import {expectType} from 'tsd';
-import {PassThrough as PassThroughStream} from 'stream';
-import ora = require('./index.js');
-import {promise} from './index.js';
+import ora, {oraPromise} from './index.js';
 
 const spinner = ora('Loading unicorns');
 ora({text: 'Loading unicorns'});
@@ -45,18 +44,18 @@ spinner.render();
 spinner.frame();
 
 const resolves = Promise.resolve(1);
-void promise(resolves, 'foo');
-void promise(resolves, {
+void oraPromise(resolves, 'foo');
+void oraPromise(resolves, {
 	stream: new PassThroughStream(),
 	text: 'foo',
 	color: 'blue',
 	isEnabled: true,
-	isSilent: false
+	isSilent: false,
 });
-void promise(async () => {
+void oraPromise(async () => {
 	await resolves;
 }, 'foo');
-void promise(async spinner => {
+void oraPromise(async spinner => {
 	spinner.prefixText = 'foo';
 	await resolves;
 	return 7;
@@ -67,5 +66,5 @@ void promise(async spinner => {
 	isEnabled: true,
 	isSilent: false,
 	successText: result => `Resolved with number ${result}`,
-	failText: 'bar'
+	failText: 'bar',
 });
